test: Consolidate and relocate test files for litellm backend and oai
